+++ src/main/java/io/github/andrewwwwwwwwwwwwwww/thp/client/ModMenuIntegration.java
@@ -1,4 +1,4 @@
-package io.github.andrewwwwwwwwwwwwwww.endbeast.client;
+package io.github.andrewwwwwwwwwwwwwww.thp.client;
 
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
@@ -6,6 +6,6 @@ import com.terraformersmc.modmenu.api.ModMenuApi;
 public class ModMenuIntegration implements ModMenuApi {
     @Override
     public ConfigScreenFactory<?> getModConfigScreenFactory() {
-        return EndBeastConfigScreen::new;
+        return ThpConfigScreen::new;
     }
 }